check_tags_function/assets/requirements.txt
google-cloud-bigquery
google-cloud-resource-manager
google-cloud-logging
google-cloud-asset
